Releases on github #197
Added release job on tags push (release will be skipped if no notes are found for a tag in the CHANGELOG.md file).
Added `contents: write` permissions to the workflow (required to create a release).
Removed --frozen-lockfile as it is automatically used on CI.
Updated versions of actions used in workflow.