Remove oldValue parameter from `notify`, as it is no longer needed
As of the previous commit, each listener has to track oldValue anyway
to avoid issues with batching, so the caller of `notify` no longer needs
to provide it.

Also removed `notify` as a publicly exposed function -- now that
listeners are automatically skipped when the value is the same as the
last call, calling `notify` with an unchanged value won't trigger the
listeners, which is the reason it was made public in the first place
(see https://github.com/orgs/nanostores/discussions/276).

A new way to solve that problem (via a custom isEqual function) will be
added in the next commit.