Don't mutate deepMap values (#305)
Previously, calling setKey on a $deepMap would mutate the value
in-place. This can cause various problems, for example #250 and #290.
Those were both solved with workarounds, however, this solves it in a
better way, by not mutating the deepMap in the first place.

Some advantages of this approach:
* No need for structuredClone
* The oldValue passed to listeners is the same object reference as the
value that was previously passed. This is how immutable stores usually
work, so it will align better with expectations, cause fewer issues, and
allow for new use cases, like the upcoming batching PR which needs that.
* Copies were already being made at each nesting level of the key being
set, they just weren't being used (so maybe this was the original
intent?). Either way, that means there's no new work being done, and
perf will actually improve since we no longer need to call
structuredClone on oldValue.

As part of this change, when a key path extends past the end of an array
the result will be a sparse array rather than filling it with undefined.
It would have required extra code (and slower perf) to keep the old
behavior, and I don't see a good reason for it. The new behavior aligns
with what JS does natively, and with other stores like SolidJS. But let
me know if that's a problem.